refactor: change package structure of dto
