+++ src/main/java/techcourse/myblog/controller/LoginController.java
@@ -11,7 +11,7 @@ import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.LoginFailException;
 import techcourse.myblog.exception.NotFoundUserException;
 import techcourse.myblog.service.LoginService;
-import techcourse.myblog.service.dto.UserDTO;
+import techcourse.myblog.service.dto.domain.UserDTO;
 
 import javax.servlet.http.HttpSession;
