+++ src/main/java/techcourse/myblog/service/dto/domain/DomainDTO.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.dto;
+package techcourse.myblog.service.dto.domain;
 
 public interface DomainDTO<T> {
     T toDomain();
