+++ src/main/java/techcourse/myblog/service/dto/domain/UserDTO.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service.dto;
+package techcourse.myblog.service.dto.domain;
 
 import techcourse.myblog.domain.User;
